@@ -1,6 +1,6 @@
 import { Location } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
@@ -16,7 +16,7 @@ import { PersonalDialogComponent } from 'src/app/shared/personal-dialog/personal
 })
 export class HotelAdminFormComponent implements OnInit {
 
-  form: FormGroup;
+  form: UntypedFormGroup;
   id: string;
   data: any = {};
 
@@ -34,18 +34,18 @@ export class HotelAdminFormComponent implements OnInit {
     private matSnackbar: MatSnackBar,
     private sharedService: SharedService
   ) {
-    this.form = new FormGroup({
-      id: new FormControl(),
-      hotelName: new FormControl(),
-      firstName: new FormControl(),
-      lastName: new FormControl(),
-      referenceName: new FormControl(),
-      email: new FormControl(),
-      mobileNo: new FormControl(),
-      address: new FormControl(),
-      country: new FormControl(),
-      province: new FormControl(),
-      status: new FormControl(),
+    this.form = new UntypedFormGroup({
+      id: new UntypedFormControl(),
+      hotelName: new UntypedFormControl(),
+      firstName: new UntypedFormControl(),
+      lastName: new UntypedFormControl(),
+      referenceName: new UntypedFormControl(),
+      email: new UntypedFormControl(),
+      mobileNo: new UntypedFormControl(),
+      address: new UntypedFormControl(),
+      country: new UntypedFormControl(),
+      province: new UntypedFormControl(),
+      status: new UntypedFormControl(),
     })
   }
 
